Separated nsfw and meme commands
